Updated EAPM blocks
